remove worker-api completely
