+++ worker/worker-api/Cargo.toml
@@ -7,15 +7,10 @@ edition = "2018"
 [dependencies]
 ws = { version = "0.7", features = ["ssl"] } 
 log = "0.4"
-hex	= "0.4"
 serde_json = "1.0"
 serde_derive = "1.0"
 sgx_crypto_helper = { rev = "v1.1.3", git = "https://github.com/apache/teaclave-sgx-sdk.git" }
 codec = { package = "parity-scale-codec", version = "1.3.1", default-features = false, features = ["derive"] }
-
-[dependencies.sp-core]
-git = "https://github.com/paritytech/substrate.git"
-rev = "a208da16"
 
 [dependencies.substratee-stf]
 path = "../../stf"
